fix: WalletConnect integration, part 4, renamed sdk files
