+++ apps/web/src/components/SendFlow/common/SingleSignPage.tsx
@@ -9,7 +9,7 @@ import { StakeSignPage } from "./StakeSignPage";
 import { TezSignPage } from "./TezSignPage";
 import { UndelegationSignPage } from "./UndelegationSignPage";
 import { UnstakeSignPage } from "./UnstakeSignPage";
-import { useSignWithBeacon } from "./useSignWithBeacon";
+import { useSignWithBeacon } from "../Beacon/useSignWithBeacon";
 
 export const SingleSignPage = (signProps: SdkSignPageProps) => {
   const operationType = signProps.operation.operations[0].type;
