+++ apps/web/src/components/SendFlow/common/StakeSignPage.tsx
