+++ apps/web/src/components/SendFlow/common/TezSignPage.test.tsx
@@ -15,7 +15,7 @@ import {
 } from "../../../testUtils";
 import { SuccessStep } from "../SuccessStep";
 import { type SdkSignPageProps, type SignHeaderProps } from "../utils";
-import { SingleSignPage } from "./BeaconSignPage";
+import { SingleSignPage } from "./SingleSignPage";
 
 jest.mock("@umami/core", () => ({
   ...jest.requireActual("@umami/core"),
