+++ apps/web/src/components/beacon/useHandleBeaconMessage.test.tsx
@@ -21,8 +21,8 @@ import { CustomError } from "@umami/utils";
 
 import { useHandleBeaconMessage } from "./useHandleBeaconMessage";
 import { act, dynamicModalContextMock, renderHook, screen, waitFor } from "../../testUtils";
-import { BatchSignPage } from "../SendFlow/Beacon/BatchSignPage";
-import { SingleSignPage } from "../SendFlow/Beacon/BeaconSignPage";
+import { BatchSignPage } from "../SendFlow/common/BatchSignPage";
+import { SingleSignPage } from "../SendFlow/common/SingleSignPage";
 import { type SdkSignPageProps, type SignHeaderProps } from "../SendFlow/utils";
 
 jest.mock("@umami/core", () => ({
