+++ apps/web/src/components/beacon/useHandleBeaconMessage.tsx
@@ -18,8 +18,8 @@ import { CustomError } from "@umami/utils";
 
 import { PermissionRequestModal } from "./PermissionRequestModal";
 import { SignPayloadRequestModal } from "./SignPayloadRequestModal";
-import { BatchSignPage } from "../SendFlow/Beacon/BatchSignPage";
-import { SingleSignPage } from "../SendFlow/Beacon/BeaconSignPage";
+import { BatchSignPage } from "../SendFlow/common/BatchSignPage";
+import { SingleSignPage } from "../SendFlow/common/SingleSignPage";
 import { type SdkSignPageProps, type SignHeaderProps } from "../SendFlow/utils";
 
 /**
